backend/requirements-rl.txt
absl-py==2.4.0
alembic==1.18.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
-e file:///home/kalxav/xdev/webwargame/CascadeProjects/windsurf-project/backend
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.4.3
cyborg @ git+https://github.com/cage-challenge/CybORG.git@2742b5e0ce4330c9b14006b38acd3b5ebe00d6fd
cycler==0.12.1
deprecated==1.3.1
distro==1.9.0
docutils==0.22.4
farama-notifications==0.0.4
fastapi==0.128.0
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-firestore==2.25.0
google-cloud-secret-manager==2.26.0
google-cloud-storage==3.10.0
google-crc32c==1.8.0
google-genai==1.67.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
greenlet==3.3.1
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
gym==0.23.1
gym-idsgame==1.0.13
gym-notices==0.1.0
gymnasium==1.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
imageio==2.37.3
iniconfig==2.3.0
invoke==2.2.1
jinja2==3.1.6
joblib==1.5.3
jsonpickle==4.1.1
kiwisolver==1.5.0
limits==5.8.0
mako==1.3.10
markdown==3.10.2
markupsafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.1
paramiko==4.0.0
passlib==1.7.4
pettingzoo==1.25.0
pillow==12.1.1
pluggy==1.6.0
prettytable==3.17.0
proto-plus==1.27.1
protobuf==6.33.6
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygame==2.6.1
pyglet==2.1.13
pygments==2.19.2
pyjwt==2.10.1
pynacl==1.6.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==82.0.1
six==1.17.0
sklearn==0.0.post12
slowapi==0.1.9
sniffio==1.3.1
sqlalchemy==2.0.46
sqlmodel==0.0.31
stable-baselines3==2.7.1
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
triton==3.6.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
werkzeug==3.1.6
wrapt==2.1.2
